feat(api): refactor subscriptions to use service users
