+++ apps/api/src/server/infrastructure/payload-builders/service-user.js
