+++ apps/api/src/server/infrastructure/payload-builders/subscriptions.js
@@ -1,0 +1,162 @@
+import { Subscription } from '@pins/applications/lib/application/subscription.js';
+import { EventType } from '@pins/event-client';
+
+/**
+ * @typedef {import('../../../message-schemas/events/nsip-subscription.d.js').NSIPSubscription} NSIPSubscription
+ * @typedef {import('@pins/applications').SubscriptionType} SubscriptionType
+ */
+
+/**
+ * Create a payload (event) for each subscription type for this subscription.
+ * Each subscription type is it's own message/event.
+ *
+ * @param {import('@pins/applications.api').Schema.Subscription} subscription
+ * @returns {NSIPSubscription[]}
+ */
+export function buildSubscriptionPayloads(subscription) {
+	/** @type {NSIPSubscription} */
+	const payload = buildSubscriptionBasePayload(subscription);
+	const types = subscriptionToTypes(subscription);
+
+	return types.map((t) => copyPayloadWithType(payload, t));
+}
+
+/**
+ * Base payload for a subscription event, without a subscription type.
+ *
+ * @param {import('@pins/applications.api').Schema.Subscription} subscription
+ * @returns {NSIPSubscription}
+ */
+export function buildSubscriptionBasePayload(subscription) {
+	/** @type {NSIPSubscription} */
+	const payload = {
+		caseReference: subscription.caseReference,
+		emailAddress: subscription.subscriber.email,
+		subscriptionType: '' // overwritten later with a real value
+	};
+	if (subscription.id) {
+		payload.subscriptionId = subscription.id;
+	}
+	if (subscription.startDate) {
+		payload.startDate = subscription.startDate.toISOString();
+	}
+
+	if (subscription.endDate) {
+		payload.endDate = subscription.endDate.toISOString();
+	}
+
+	if (subscription.language) {
+		payload.language = subscription.language;
+	}
+
+	return payload;
+}
+
+/**
+ * @param {import('@pins/applications.api').Schema.Subscription} subscription
+ * @returns {SubscriptionType[]}
+ */
+export function subscriptionToTypes(subscription) {
+	/** @type {SubscriptionType[]} */
+	const types = [];
+
+	if (subscription.subscribedToAllUpdates) {
+		types.push(Subscription.Type.allUpdates);
+	} else {
+		if (subscription.subscribedToApplicationDecided) {
+			types.push(Subscription.Type.applicationDecided);
+		}
+		if (subscription.subscribedToApplicationSubmitted) {
+			types.push(Subscription.Type.applicationSubmitted);
+		}
+		if (subscription.subscribedToRegistrationOpen) {
+			types.push(Subscription.Type.registrationOpen);
+		}
+	}
+	return types;
+}
+
+/**
+ *
+ * @param {SubscriptionType[]} types
+ * @param {import('@prisma/client').Prisma.SubscriptionCreateInput} subscription
+ */
+export function typesToSubscription(types, subscription) {
+	// ensure all fields are reset to false (not just left as they are) for updates
+	subscription.subscribedToAllUpdates = false;
+	subscription.subscribedToApplicationDecided = false;
+	subscription.subscribedToApplicationSubmitted = false;
+	subscription.subscribedToRegistrationOpen = false;
+
+	if (types.includes(Subscription.Type.allUpdates)) {
+		subscription.subscribedToAllUpdates = true;
+	} else {
+		for (const type of types) {
+			switch (type) {
+				case Subscription.Type.applicationSubmitted:
+					subscription.subscribedToApplicationSubmitted = true;
+					break;
+				case Subscription.Type.applicationDecided:
+					subscription.subscribedToApplicationDecided = true;
+					break;
+				case Subscription.Type.registrationOpen:
+					subscription.subscribedToRegistrationOpen = true;
+					break;
+			}
+		}
+	}
+}
+
+/**
+ * @typedef {Object<string,NSIPSubscription[]>} EventsByType
+ */
+
+/**
+ * Given an existing and updated subscription, what are the subscription type changes?
+ * For each of those changes, generate an associated subscription event
+ *
+ * For example, if a subscription is updated to add an "applicationDecided" subscription,
+ * a Create event will be sent for that, and an Update event for any existing subscriptions.
+ *
+ * @param {import('@pins/applications.api').Schema.Subscription} existingSub
+ * @param {import('@pins/applications.api').Schema.Subscription} newSub
+ * @returns {EventsByType}
+ */
+export function subscriptionTypeChanges(existingSub, newSub) {
+	const existingTypes = subscriptionToTypes(existingSub);
+	const newTypes = subscriptionToTypes(newSub);
+	const basePayload = buildSubscriptionBasePayload(newSub);
+
+	const addedTypes = newTypes.filter((t) => !existingTypes.includes(t));
+	const removedTypes = existingTypes.filter((t) => !newTypes.includes(t));
+	const unchangedTypes = newTypes.filter((t) => existingTypes.includes(t));
+
+	/** @type {EventsByType} */
+	const byType = {};
+
+	if (addedTypes.length > 0) {
+		byType[EventType.Create] = addedTypes.map((t) => copyPayloadWithType(basePayload, t));
+	}
+	if (unchangedTypes.length > 0) {
+		byType[EventType.Update] = unchangedTypes.map((t) => copyPayloadWithType(basePayload, t));
+	}
+	if (removedTypes.length > 0) {
+		byType[EventType.Delete] = removedTypes.map((t) => copyPayloadWithType(basePayload, t));
+	}
+
+	return byType;
+}
+
+/**
+ * Base payload for a subscription event, without a subscription type.
+ *
+ * @param {NSIPSubscription} basePayload
+ * @param {SubscriptionType} type
+ * @returns {NSIPSubscription}
+ */
+function copyPayloadWithType(basePayload, type) {
+	return {
+		...basePayload,
+		subscriptionType: type
+	};
+}
